feat: adicionar página de solicitações de adoção recebidas
- Criada a página de solicitações de adoção recebidas com componentes para aprovação e rejeição
- Implementada lógica para carregar e gerenciar solicitações
- Atualizado o arquivo de rotas para incluir a nova página
- Ajustadas as configurações do docker para ambiente de desenvolvimento
