+++ docker-compose.dev.yml
@@ -3,7 +3,7 @@ services:
     image: node:24.12.0-alpine
     working_dir: /app
     environment:
-      VITE_AUTH_STRATEGY: jwt
+      VITE_AUTH_STRATEGY: mock
       VITE_API_URL: http://localhost:3000/api
     ports:
       - "5173:5173"
